Modularize table icons constant.
