+++ src/Components/AllCandidates/AllCandidates.jsx
@@ -32,10 +32,8 @@ const Wrapper = styled.div`
     font-weight: normal;
     font-size: 24px;
     line-height: 36px;
-    max-width: 854px;
     width: 90vw;
-    margin: 0 auto;
-    padding-bottom: 20px;
+    margin: 0;
     text-align: left;
   }
   .table {
