+++ src/Components/AllCandidates/CommitteeReview.jsx
@@ -37,8 +37,7 @@ const Wrapper = styled.div`
     line-height: 36px;
     max-width: 854px;
     width: 90vw;
-    margin: 0 auto;
-    padding-bottom: 20px;
+    margin: 0;
   }
   p {
     font-family: Roboto;
@@ -50,7 +49,7 @@ const Wrapper = styled.div`
     color: #2261ad;
     max-width: 854px;
     width: 90vw;
-    margin: 0 auto;
+    margin: 0;
     padding-bottom: 20px;
     vertical-align: center;
   }
@@ -152,12 +151,13 @@ class CommitteeReview extends Component {
           <>
             <Header>
               <p align="left" onClick={this.goBack}>
-                <span style={{ cursor: "pointer" }}>
+                <span style={{ cursor: "pointer", color: '#2261AD' }}>
                   <FontAwesomeIcon
                     style={{
                       height: "25px",
                       width: "25px",
-                      verticalAlign: "-0.5em"
+                      verticalAlign: "-0.5em",
+                      color: '#2261AD'
                     }}
                     icon={faAngleLeft}
                   />
